feat(servidor): agregar el trait HayRelacionUsuarioYCliente
La comprobación va a ser utilizada en todas las políticas de los modelos relacionados con cliente
